feat(lmdb): fix code
